requirements.txt

# Langchain core
langchain
langchain-core

#Google Gemini Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

streamlit

pillow
ipython